Update email.html
Found a hack to make full-width hero responsive in Gmail desktop and mobile (Android & iOS app).

Set height attribute in the main <img> tag to auto; and declare max-height in the img's <style> tag. Something about the order in which these are red now makes it fluid in Gmail and doesn't break it in Outlook (as height="auto" often does in standard coding).